+++ src/modules/AlephVaultSettlement.sol
@@ -82,7 +82,7 @@ contract AlephVaultSettlement is IAlephVaultSettlement, AlephVaultBase {
         if (_settlementParams.toBatchId <= _depositSettleId) {
             revert NoDepositsToSettle();
         }
-        uint8 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
+        uint32 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
         _validateNewTotalAssets(_shareClass.shareSeriesId, _lastConsolidatedSeriesId, _settlementParams.newTotalAssets);
         if (_sd.isSettlementAuthEnabled) {
             AuthLibrary.verifySettlementAuthSignature(
@@ -103,7 +103,7 @@ contract AlephVaultSettlement is IAlephVaultSettlement, AlephVaultBase {
             _settlementParams.toBatchId,
             _settlementParams.newTotalAssets
         );
-        uint8 _settlementSeriesId =
+        uint32 _settlementSeriesId =
             _handleSeriesAccounting(_shareClass, _settlementParams.classId, _settlementParams.toBatchId);
         IAlephVault.ShareSeries storage _shareSeries = _shareClass.shareSeries[_settlementSeriesId];
         SettleDepositDetails memory _settleDepositDetails = SettleDepositDetails({
@@ -232,7 +232,7 @@ contract AlephVaultSettlement is IAlephVaultSettlement, AlephVaultBase {
         if (_settleUptoBatchId <= _redeemSettleId) {
             revert NoRedeemsToSettle();
         }
-        uint8 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
+        uint32 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
         _validateNewTotalAssets(_shareClass.shareSeriesId, _lastConsolidatedSeriesId, _settlementParams.newTotalAssets);
         if (_sd.isSettlementAuthEnabled) {
             AuthLibrary.verifySettlementAuthSignature(
@@ -372,12 +372,12 @@ contract AlephVaultSettlement is IAlephVaultSettlement, AlephVaultBase {
      */
     function _handleSeriesAccounting(IAlephVault.ShareClass storage _shareClass, uint8 _classId, uint48 _toBatchId)
         internal
-        returns (uint8 _seriesId)
+        returns (uint32 _seriesId)
     {
         // for non-incentive classes, all settlements take place in the lead series
         if (_shareClass.shareClassParams.performanceFee > 0) {
-            uint8 _shareSeriesId = _shareClass.shareSeriesId;
-            uint8 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
+            uint32 _shareSeriesId = _shareClass.shareSeriesId;
+            uint32 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
             // if new lead series highwatermark is not reached, deposit settlements must take place in a new series
             // if a new highwater mark is reached in this cycle, it will be updated in _accumalateFees function
             // hence, after fee accumalation process, the lead highwater mark is either greater than or equal to the lead price per share
@@ -401,8 +401,8 @@ contract AlephVaultSettlement is IAlephVaultSettlement, AlephVaultBase {
      * @param _newTotalAssets The new total assets after settlement.
      */
     function _validateNewTotalAssets(
-        uint8 _shareSeriesId,
-        uint8 _lastConsolidatedSeriesId,
+        uint32 _shareSeriesId,
+        uint32 _lastConsolidatedSeriesId,
         uint256[] calldata _newTotalAssets
     ) internal pure {
         if (_newTotalAssets.length != _shareSeriesId - _lastConsolidatedSeriesId + 1) {
@@ -421,14 +421,14 @@ contract AlephVaultSettlement is IAlephVaultSettlement, AlephVaultBase {
     function _accumulateFees(
         IAlephVault.ShareClass storage _shareClass,
         uint8 _classId,
-        uint8 _lastConsolidatedSeriesId,
+        uint32 _lastConsolidatedSeriesId,
         uint48 _toBatchId,
         uint256[] calldata _newTotalAssets
     ) internal {
         uint48 _lastFeePaidId = _shareClass.lastFeePaidId;
         if (_toBatchId > _lastFeePaidId) {
-            for (uint8 _i = 0; _i < _newTotalAssets.length; _i++) {
-                uint8 _seriesId = _i > SeriesAccounting.LEAD_SERIES_ID
+            for (uint32 _i = 0; _i < _newTotalAssets.length; _i++) {
+                uint32 _seriesId = _i > SeriesAccounting.LEAD_SERIES_ID
                     ? _lastConsolidatedSeriesId + _i
                     : SeriesAccounting.LEAD_SERIES_ID;
                 // update the series total assets and shares
@@ -462,7 +462,7 @@ contract AlephVaultSettlement is IAlephVaultSettlement, AlephVaultBase {
         uint48 _toBatchId,
         uint48 _lastFeePaidId,
         uint8 _classId,
-        uint8 _seriesId
+        uint32 _seriesId
     ) internal returns (uint256) {
         if (_newTotalAssets == 0) {
             return 0;
